% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bas_clu.R
\name{basclu}
\alias{basclu}
\title{Run BasClu.}
\usage{
basclu(x, s, NT, nthin, nupd, Nburn, clustering = "BasCluZ", ...)
}
\arguments{
\item{x}{Log-transformed count matrix}

\item{s}{Log library sizes}

\item{NT}{Number of MCMC iterations}

\item{nthin}{For thinning: keep sample every nthin iteration}

\item{nupd}{For updating random walk during burn-in}

\item{Nburn}{Number of iterations for burn-in}

\item{clustering}{Clustering method to use}

\item{...}{Additional params}
}
\value{
z Z-matrix
}
\description{
Run BasClu.
}
